pulled subproc tests to a separate file
